Made a simple menu and added interaction to the buttons to improve the user experience, and will continue to be updated in the future.
I built a menu with buttons linking to each member's page. And on these buttons, in order for the user to know more clearly whether they clicked the button, these buttons will give the user a feedback. When the user swipes over the button, the button changes color, and it's a more noticeable color change. At the same time, a character representing our three girls was added to the page, and while she may not be the protagonist of our game, she is our representative, which is fun! At the same time, the background color is added to better divide the area. This time the menu function is not perfect, and I will continue to modify it next time.