This is the  menu I updated.
I added a button sound for the button on this page, it is a very cute sound. Also added brisk background music. Have a relaxing effect. At the same time, the color of the button is designed to change after clicking, the position of some details is changed, and the title is added. I will make the next changes.